Revert "Revert commonization of `inline fun <reified VM> viewModel(..)`"
This reverts commit dc44b8489c3c03ad5aa7d25be3bda841cf81c93c.
